Mma 7160 ytelsesforbedringer (#44)
* optimalisering av dokarkivavlevering

* Bedre feilhåndtering om vi ikke får konvertert dokumentet

* batchsize ligger i vault, fjern fra application.properties

* Remove unused camel properties

* Replace camelLoopIndex med camelSplitIndex

+++ app/src/main/java/no/nav/dokarkivavlevering/avlevering/config/AvleveringProperties.java
@@ -27,7 +27,7 @@ public class AvleveringProperties {
 	String tema;
 	@NotNull
 	@Max(1000) // max IN query i Oracle
-	Long batchsize;
+	int batchsize;
 	@NotEmpty
 	String ststokenurl;
 	@NotEmpty
